requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
gunicorn==20.1.0
idna==3.3
paystackapi==2.1.0
Pillow==9.1.0
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
